+++ apps/codeforafrica/src/components/MobileNavBar/MobileNavBar.snap.js
