+++ apps/codeforafrica/src/components/NavBar/NavBar.snap.js
