+++ apps/codeforafrica/src/components/NavBar/NavBar.test.js
