+++ apps/codeforafrica/src/components/NavBar/index.js
@@ -1,44 +1,3 @@
-import { NavBar as NavigationBar, Section } from "@commons-ui/core";
-import PropTypes from "prop-types";
-import React from "react";
-
-import DesktopNavBar from "@/codeforafrica/components/DesktopNavBar";
-import MobileNavBar from "@/codeforafrica/components/MobileNavBar";
-
-function NavBar({ menu }) {
-  return (
-    <NavigationBar>
-      <Section sx={{ px: { xs: "20px", sm: 0 } }}>
-        <MobileNavBar
-          menu={menu}
-          direction="column"
-          sx={{
-            display: { xs: "flex", md: "none" },
-          }}
-        />
-        <DesktopNavBar
-          menu={menu}
-          direction="row"
-          sx={{
-            display: { xs: "none", md: "flex" },
-          }}
-        />
-      </Section>
-    </NavigationBar>
-  );
-}
-
-NavBar.propTypes = {
-  menu: PropTypes.arrayOf(
-    PropTypes.shape({
-      label: PropTypes.string,
-      href: PropTypes.string,
-    })
-  ),
-};
-
-NavBar.defaultProps = {
-  menu: undefined,
-};
+import NavBar from "./NavBar";
 
 export default NavBar;
